requirements.txt
dnspython>=2.7.0
paho-mqtt>=2.1.0
psycopg2>=2.9.10
python-dotenv>=1.1.0